Split server types into domain modules
