+++ AGENTS.md
@@ -78,6 +78,7 @@ Follow ADR-0051 for shared contracts, logging, and ownership boundaries.
 - For Workspace Context work, keep `src/services/workspaceContext/service.ts` as the coordinator. Keep markdown storage, instruction management, source planning, processor prompting, run state, and scheduling inside focused modules under `src/services/workspaceContext/` as the feature grows; do not reintroduce graph, candidate-review, MCP subsystems, or the old Memory Review scheduler/UI workflow unless a new ADR supersedes the markdown-first design. Workspace Context maintenance is the automatic path that consumes accepted active Memory inbox entries and deletes them after folding them into context markdown.
 - For Workspace Routines work, keep Agent Cockpit as the infrastructure layer under `src/services/routines/`: authoring contracts, proposal validation, scheduling, run folders/history, and outreach. Keep task-specific intelligence in routine markdown executed by the selected harness, and keep shared route/client validators in `src/contracts/routines.ts`. Routines are opt-in per workspace through `WorkspaceIndex.routinesEnabled`; disabled workspaces must not install managed `AGENTS.md` instructions, create routine scaffolding from passive reads, or run scheduled routines.
 - Put shared request/response shapes and runtime validators in `src/contracts/` whenever routes, clients, tests, or specs need the same boundary. Contract files imported by web or mobile must stay browser-safe and must not import server-only modules.
+- Server type definitions live in declaration-only domain modules under `src/types/` and are re-exported through `src/types/index.ts`; contracts keep importing exactly `../types`, and new server-only types should go in the matching domain file rather than the barrel.
 - Keep browser-visible stream frames in `src/contracts/streamFrames.ts`. Backend adapters may use broader server-only `StreamEvent` shapes, but web/mobile clients should consume only the browser-safe frame union. Desktop stream-frame state changes belong in the pure reducer under `web/AgentCockpitWeb/src/stream/`; keep `streamStore.js` as the socket/API/timer effect adapter.
 - Web and mobile clients should import shared types only from browser-safe contracts. Do not reach into backend services, route modules, filesystem helpers, or server-only types from frontend code.
 - Keep large UI entrypoints as composition layers. Move pure projection, parsing, state-provider, viewport, and formatting behavior into focused modules or hooks with direct tests.
