+++ docs/spec-backend-services.md
@@ -12,7 +12,7 @@
 
 `ChatService` remains the public facade for conversation, workspace, memory, KB, Workspace Context, settings, queue, usage, artifact, workspace-archive, workspace-location, workspace-instruction, and worktree-isolation operations. Pure attachment helpers stay in `src/services/chat/attachments.ts` and are re-exported where legacy imports still expect them from `chatService.ts`; generated artifact persistence lives in `src/services/chat/artifactStore.ts` behind `createConversationArtifact`; workspace/session pathing, `index.json` and `session-N.json` reads/writes, and the convId→workspaceId lookup rebuild live in `src/services/chat/workspaceSessionStore.ts`; conversation sidebar listing and simple lifecycle/index mutations (`rename`, archive/restore, unread state, external session ID persistence) live in `src/services/chat/conversationLifecycleStore.ts`; message append/edit/pin operations, session history/message reads, and message-window/pinned-message projection helpers live in `src/services/chat/conversationMessageStore.ts`; conversation/session usage accumulation and per-backend usage buckets live in `src/services/chat/conversationUsageStore.ts`; workspace Memory pathing, snapshot/state JSON persistence, and consolidation audit files live in `src/services/chat/workspaceMemoryStore.ts`; workspace KB pathing plus SQLite DB and PGLite vector-store caches live in `src/services/chat/workspaceKnowledgeStore.ts`; message-queue persistence and legacy queue normalization live in `src/services/chat/messageQueueStore.ts` behind the same public `getQueue`/`setQueue`/`clearQueue` methods; workspace archive summary/lifecycle persistence lives in `src/services/chat/workspaceArchiveStore.ts`; workspace file snapshots, checksum verification, safe extraction, product-trash cleanup, and default restore destinations live in `src/services/chat/workspaceSnapshotService.ts`; usage-ledger persistence/aggregation lives in `src/services/chat/usageLedgerStore.ts` behind `addUsage`, `getUsageStats`, and `clearUsageStats`; external Claude transcript usage import lives in `src/services/claudeTranscriptUsageImportService.ts` and feeds only the global usage ledger; usage-pricing defaults, estimation, and mutable overrides live in `src/services/usagePricing/` behind `getUsagePricingCatalog`, `saveUsagePricingOverrides`, and `clearUsagePricingOverrides`; stable workspace identity and legacy storage-key resolution live in `src/services/chat/workspaceIdentityStore.ts`; workspace KB, Workspace Context, and Workspace Routines enablement/settings flags live in `src/services/chat/workspaceFeatureSettingsStore.ts`; workspace-instruction compatibility detection and pointer-file creation live in `src/services/chat/workspaceInstructionStore.ts` behind the existing instruction facade methods; Git worktree probing/lifecycle commands live in `src/services/chat/worktreeIsolationService.ts` behind ChatService orchestration. This keeps helper/store behavior testable without changing the service API.
 
-Chat API route registration is split by domain. `src/routes/chat.ts` owns dependency construction, stream lifecycle orchestration, and route composition. Focused routers own status/settings (`statusRoutes.ts`), CLI profile metadata/auth (`cliProfileRoutes.ts`), conversation/session/queue CRUD (`conversationRoutes.ts`), stream send/input/abort/status (`streamRoutes.ts`), backend goal mutations (`goalRoutes.ts`), conversation uploads and workspace file delivery (`uploadRoutes.ts`), local directory picker routes (`filesystemRoutes.ts`), workspace archive routes (`workspaceArchiveRoutes.ts`), workspace location routes (`workspaceLocationRoutes.ts`), workspace instruction routes (`workspaceInstructionRoutes.ts`), workspace file explorer routes (`explorerRoutes.ts`), read-only workspace/conversation Git status/diff routes (`gitRoutes.ts`), workspace worktree-isolation status/toggle routes (`worktreeIsolationRoutes.ts`), memory routes (`memoryRoutes.ts`), Workspace Context routes including conversation-scoped context/reference/asset previews (`workspaceContextRoutes.ts`), Workspace Routines routes (`routineRoutes.ts`), and KB routes (`kbRoutes.ts`). Common Express helpers live in `src/routes/chat/routeUtils.ts`; goal lifecycle transcript formatting and runtime snapshots live in `src/services/chat/goalEventMessages.ts` so routes and stream processing persist the same `Message.goalEvent` shape. Runtime request validators and small shared response contracts live in focused files under `src/contracts/`; browser-safe wire types avoid importing server-only `src/types` where web/mobile clients use type-only references.
+Chat API route registration is split by domain. `src/routes/chat.ts` owns dependency construction, stream lifecycle orchestration, and route composition. Focused routers own status/settings (`statusRoutes.ts`), CLI profile metadata/auth (`cliProfileRoutes.ts`), conversation/session/queue CRUD (`conversationRoutes.ts`), stream send/input/abort/status (`streamRoutes.ts`), backend goal mutations (`goalRoutes.ts`), conversation uploads and workspace file delivery (`uploadRoutes.ts`), local directory picker routes (`filesystemRoutes.ts`), workspace archive routes (`workspaceArchiveRoutes.ts`), workspace location routes (`workspaceLocationRoutes.ts`), workspace instruction routes (`workspaceInstructionRoutes.ts`), workspace file explorer routes (`explorerRoutes.ts`), read-only workspace/conversation Git status/diff routes (`gitRoutes.ts`), workspace worktree-isolation status/toggle routes (`worktreeIsolationRoutes.ts`), memory routes (`memoryRoutes.ts`), Workspace Context routes including conversation-scoped context/reference/asset previews (`workspaceContextRoutes.ts`), Workspace Routines routes (`routineRoutes.ts`), and KB routes (`kbRoutes.ts`). Common Express helpers live in `src/routes/chat/routeUtils.ts`; goal lifecycle transcript formatting and runtime snapshots live in `src/services/chat/goalEventMessages.ts` so routes and stream processing persist the same `Message.goalEvent` shape. Runtime request validators and small shared response contracts live in focused files under `src/contracts/`; browser-safe wire types avoid importing server-only `src/types` where web/mobile clients use type-only references. Server-only type declarations live in declaration-only domain modules under `src/types/` and are re-exported through `src/types/index.ts`; contracts may import only that `../types` barrel, not deep type modules.
 
 Conversation upload image normalization is owned by focused chat modules rather than `uploadRoutes.ts`. `src/services/chat/dngPreview.ts` is a pure TIFF/DNG parser for the supported baseline TIFF layout (`II`/`MM` endian markers, TIFF magic 42, 32-bit IFD offsets). It scans IFD0, linked IFDs, and SubIFDs up to a bounded depth, considers only single-strip JPEG preview candidates (`Compression=7` or JPEG SOI bytes) with non-linear-raw photometric interpretation and 8-bit-or-smaller samples, parses JPEG SOF dimensions without fully decoding pixels, trims at JPEG EOI when present, and returns the largest readable candidate. Linear raw subimages (`PhotometricInterpretation=34892`) are intentionally skipped; Agent Cockpit does not demosaic RAW sensor data for chat uploads. `src/services/chat/uploadImageNormalization.ts` applies this parser only to `.dng` uploads, writes `<original>.preview.jpg`, and caps the generated JPEG to `CHAT_UPLOAD_IMAGE_MAX_LONG_EDGE_PX = 2576` with `@napi-rs/canvas`/JPEG quality 92 when the embedded preview exceeds the cap. Other upload formats pass through unchanged. The generated JPEG sidecar is the only `AttachmentMeta` returned to clients for the upload, while the original DNG remains on disk until the sidecar is deleted. This decision is captured in [ADR-0083](adr/0083-extract-dng-embedded-previews-for-chat-uploads.md).
